+++ app/src/main/AndroidManifest.xml
@@ -20,44 +20,44 @@
         android:theme="@style/AppTheme"
         tools:ignore="LockedOrientationActivity">
         <activity
-            android:name=".addfood.FoodSearchActivity"
-            android:parentActivityName=".addfood.AddFoodActivity"
+            android:name="com.bruhascended.fitapp.ui.addfood.FoodSearchActivity"
+            android:parentActivityName="com.bruhascended.fitapp.ui.addfood.AddFoodActivity"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustPan" />
         <activity
-            android:name=".logweight.LogWeightActivity"
+            android:name="com.bruhascended.fitapp.ui.logweight.LogWeightActivity"
             android:label="@string/log_weight"
             android:screenOrientation="portrait">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.bruhascended.fitapp.main.MainActivity" />
+                android:value="com.bruhascended.fitapp.ui.main.MainActivity" />
         </activity>
         <activity
-            android:name=".addworkout.AddWorkoutActivity"
+            android:name="com.bruhascended.fitapp.ui.addworkout.AddWorkoutActivity"
             android:label="@string/add_new_workout"
             android:screenOrientation="portrait">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.bruhascended.fitapp.main.MainActivity" />
+                android:value="com.bruhascended.fitapp.ui.main.MainActivity" />
         </activity>
         <activity
-            android:name=".addfood.AddFoodActivity"
+            android:name="com.bruhascended.fitapp.ui.addfood.AddFoodActivity"
             android:label="@string/add_new_food_entry"
             android:screenOrientation="portrait">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.bruhascended.fitapp.main.MainActivity" />
+                android:value="com.bruhascended.fitapp.ui.main.MainActivity" />
         </activity>
         <activity
-            android:name=".capturefood.CaptureFoodActivity"
+            android:name="com.bruhascended.fitapp.ui.capturefood.CaptureFoodActivity"
             android:screenOrientation="portrait"
             android:label="@string/capture_a_food_item">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.bruhascended.fitapp.main.MainActivity" />
+                android:value="com.bruhascended.fitapp.ui.main.MainActivity" />
         </activity>
         <activity
-            android:name=".main.MainActivity"
+            android:name="com.bruhascended.fitapp.ui.main.MainActivity"
             android:launchMode="singleTask"
             android:screenOrientation="portrait">
             <intent-filter>
