+++ app/src/main/java/com/bruhascended/fitapp/ui/addfood/FoodSearchActivity.kt
@@ -1,4 +1,4 @@
-package com.bruhascended.fitapp.addfood
+package com.bruhascended.fitapp.ui.addfood
 
 import android.app.SearchManager
 import android.os.Bundle
@@ -12,9 +12,9 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.bruhascended.fitapp.R
 import com.bruhascended.fitapp.databinding.ActivityFoodSearchBinding
-import com.bruhascended.fitapp.util.setupToolbar
+import com.bruhascended.fitapp.ui.util.setupToolbar
 import com.bruhascended.api.models.foods.Food
-import com.bruhascended.fitapp.capturefood.PredictionPresenter
+import com.bruhascended.fitapp.ui.capturefood.PredictionPresenter
 import kotlinx.coroutines.*
 
 class FoodSearchActivity : AppCompatActivity() {
