+++ app/src/main/java/com/bruhascended/fitapp/ui/util/HideFabOnScrollBehavior.kt
@@ -1,4 +1,4 @@
-package com.bruhascended.fitapp.util
+package com.bruhascended.fitapp.ui.util
 
 import android.animation.Animator
 import android.animation.AnimatorListenerAdapter
