+++ app/src/main/res/layout/activity_add_food.xml
@@ -7,7 +7,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="?backgroundColor"
-        tools:context=".addfood.AddFoodActivity">
+        tools:context=".ui.addfood.AddFoodActivity">
 
         <com.google.android.material.appbar.AppBarLayout
             android:layout_width="match_parent"
