+++ app/src/main/res/layout/activity_add_workout.xml
@@ -6,7 +6,7 @@
     <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".addworkout.AddWorkoutActivity">
+        tools:context=".ui.addworkout.AddWorkoutActivity">
 
         <com.google.android.material.appbar.AppBarLayout
             android:layout_width="match_parent"
