+++ app/src/main/res/layout/activity_food_search.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".addfood.FoodSearchActivity">
+    tools:context=".ui.addfood.FoodSearchActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
