+++ app/src/main/res/navigation/nav_graph_bottom.xml
@@ -7,17 +7,17 @@
 
     <fragment
         android:id="@+id/dashboardFragment"
-        android:name="com.bruhascended.fitapp.dashboard.DashboardFragment"
+        android:name="com.bruhascended.fitapp.ui.dashboard.DashboardFragment"
         android:label="@string/dashboard"
         tools:layout="@layout/fragment_dashboard" />
     <fragment
         android:id="@+id/challengesFragment"
-        android:name="com.bruhascended.fitapp.challenges.ChallengesFragment"
+        android:name="com.bruhascended.fitapp.ui.challenges.ChallengesFragment"
         android:label="@string/challenges"
         tools:layout="@layout/fragment_challenges" />
     <fragment
         android:id="@+id/journalFragment"
-        android:name="com.bruhascended.fitapp.journal.JournalFragment"
+        android:name="com.bruhascended.fitapp.ui.journal.JournalFragment"
         android:label="@string/journal"
         tools:layout="@layout/fragment_journal" />
 </navigation>
